pyproject.toml
[tool.poetry]
name = "zeromq-patterns"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12"
pyzmq = "^25.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
